feat(signup): create signup page
